// client/src/components/raveStream/swipe/SwipeStream.enum.tsx
/**
 * SwipeStream.enum.tsx
 *
 * Enumerators for the swipe streams.
 */

/**
 * View types for the swipe stream.
 */
export enum SwipeView {
  PRODUCT = 'product',
  REVIEW = 'review',
  VIDEO = 'video'
}

/**
 * Positioning types for the video overlay.
 */
export enum VideoPosition {
  SHOWING = 'showing',
  NEXT = 'next',
  PREVIOUS = 'previous'
}